isolate multi_view_reconstruction with k-neighbours in a specific module (for compatibility) redefine ImageView with simpler attributes. Keep OldImageView for use with k-neighbours reconstruction add multi_tolerant_reconstruction algorithm
