+++ src/openalea/phenomenal/multi_view_reconstruction/__init__.py
@@ -29,6 +29,7 @@ Main's function
 
 from .multi_view_reconstruction import *
 from ._multi_view_reconstruction_octree import *
+from ._multi_view_reconstruction_neighbours import *
 # ==============================================================================
 
 __all__ = [s for s in dir() if not s.startswith("_")]
