+++ src/openalea/phenomenal/object/__init__.py
@@ -29,7 +29,7 @@ Data Structure
 # ==============================================================================
 
 
-from .imageView import ImageView
+from .imageView import ImageView, OldImageView
 from .image3D import Image3D
 from .voxelOctree import VoxelOctree
 from .voxelGrid import VoxelGrid
